Hide the Drawable class. Drawable class holds tgfx::Device pointer to prevent repeated creation
