+++ include/pag/pag.h
@@ -28,13 +28,14 @@ namespace tgfx {
 struct Rect;
 class Context;
 class Surface;
-class Device;
 }  // namespace tgfx
 
 namespace pag {
 class Recorder;
 
 class RenderCache;
+
+class Drawable;
 
 class Content {
  public:
@@ -1116,57 +1117,6 @@ class PAG_API PAGFile : public PAGComposition {
 class Composition;
 
 class PAGPlayer;
-
-class Drawable {
- public:
-  virtual ~Drawable() = default;
-
-  /**
-   * Returns the width of this drawable.
-   */
-  virtual int width() const = 0;
-
-  /**
-   * Returns the height of this drawable.
-   */
-  virtual int height() const = 0;
-
-  /**
-   * Update the size of the drawable, and reset the associated backend render target.
-   */
-  virtual void updateSize() = 0;
-
-  /**
-   * Creates a new Surface from this drawable.
-   */
-  virtual std::shared_ptr<tgfx::Surface> createSurface(tgfx::Context* context) = 0;
-
-  /**
-   * Apply all pending changes to the drawable.
-   * Note: The associated GPUDevice must be the current rendering device on the calling thread.
-   */
-  virtual void present(tgfx::Context* context) = 0;
-
-  /**
-   * Locks the rendering context associated with this drawable, if another thread has already locked
-   * the drawable by lockContext(), a call to lockContext() will block execution until the context
-   * is available. The returned context can be used to draw graphics. A nullptr is returned If the
-   * context can not be locked on the calling thread, and leaves the drawable unlocked.
-   */
-  virtual tgfx::Context* lockContext() = 0;
-
-  /**
-   * Unlocks the context. After this method is called all subsequent calls on the previously returned
-   * context will be invalid and may lead to runtime crash.
-   */
-  virtual void unlockContext() = 0;
-
-  /**
-   * Set the presenting timeStamp, used for android
-   */
-  virtual void setTimeStamp(int64_t) {
-  }
-};
 
 class GLRestorer;
