+++ src/rendering/PAGPlayer.cpp
@@ -19,7 +19,7 @@
 #include "base/utils/TimeUtil.h"
 #include "core/Clock.h"
 #include "pag/file.h"
-#include "pag/pag.h"
+#include "rendering/Drawable.h"
 #include "rendering/FileReporter.h"
 #include "rendering/caches/RenderCache.h"
 #include "rendering/layers/PAGStage.h"
